+++ ecommerce/src/main/java/com/project/ecommerce/service/IProductService.java
@@ -1,7 +1,6 @@
 package com.project.ecommerce.service;
 
 import com.project.ecommerce.dto.CategoryDto;
-import com.project.ecommerce.dto.ProductDto;
 import com.project.ecommerce.dto.SubCategoryDto;
 import com.project.ecommerce.form.ProductForm;
 
@@ -12,7 +11,7 @@ public interface IProductService {
     List<CategoryDto> getAllCategory();
     List<SubCategoryDto> getALLSubCategory();
     void updateProduct(ProductForm productForm);
-    void deleteProduct(Integer productId);
+    void deleteProduct(Long productId);
     List<ProductForm> getAllProductByVendorId(Long vendorId);
-    ProductForm getVendorProduct(Integer productId);
+    ProductForm getVendorProduct(Long productId);
 }
